Valencioo Hoffman - De enemy robot AI heb ik veranderd de enemies kunnen nu dekking zoeken en ze hebben een manoeuvre om te ontwijken wanneer ze aan het aanvallen zijn. Ook heb ik particles toegevoegd zodat je feedback krijgt wanneer je een powerup pakt. Tot slot heb ik de cameracontroller script aangepast waarbij de camera niet meer door de muur heen gaat.
Nu ga ik aan de slag met de enemy spawners ( ook power ups spawners )

+++ Game/Super Custom Robot Arena/Assets/Scripts/Enemy/EnemyHead.cs
@@ -96,8 +96,8 @@ public class EnemyHead : Part {
 	/// Heals the armor.
 	/// </summary>
 	/// <param name="h">Health.</param>
-	public override void ArmorHeal(int h){
-		this.mArmorHealth += (this.mArmorHealth / h);
+	public override void ArmorHeal(double h){
+		this.mArmorHealth += (float)(this.mArmorHealth / h);
 		this.UpdateShieldBar();
 	}
